Set dialog status using ajax instead
The previous solution with a big form was very easy to get wrong by
accident, such as setting back the status of a speaker who confirmed
their talk while the form was being edited.

Instead, use ajax and set the status individually for each talk. The
only remaining race condition then is if two admins change the status of
the same talk, and in that case, let's solve it in meatspace.

Keep the restriction to never set a talk directly to approved, of
course, as that is just preventing an unnecessary user error.