Switch font for user data on invoices
This switches invoices to use DejaVu Serif font instead of Times,
because the Times font does not have all characters we need. In
particular this has become obvious with Turkish characters.

We use DjeaVu Serif because that's what we already use for conference
badges and PDF schedules so we know it exists on our servers and that it
works for those. The DejaVu font is slightly bigger than Times, so
decrease the size by one point to make it look roughly the same as
before.