Fix balance queries to work when previous year is not closed
The same bug existed in the paypal and the trustly verifiers, in that
they assumed the previous year had been closed. Clearly they have never
been run in january before -- simply because they did not exist last
january.

Instead do some SQL magic to traverse the history of years and support
multiple open years while at it. Centralize this query to the accounting
module as basically the same query was used in more than one place.