Add JSON output format for attendee reports
This will be used together with the new badge rendering framework, but
can also be used indepdently.

To implement this, add concept of a "safe export" of registrations that
generate a dict structure (that is later exported as json) with the data
from a conference registration and attached objects that are safe to
export this way.