Add support for rendering badges using jinja/json
Instead of having to write badges as python code (with all the
implications that had - such as not letting outsiders have access
to them for security reasons), they can now be written in JSON using
Jinja templates to insert replacement variables and write simpler logic.

JSON based badges are loaded from badge.json in the directory containing
the conference jinja templates. Images used are all loaded from the
static directory in the same enforced structure.

If Jinja templates are enabled for conference pages, they are
automatically enabled for badges as well - there is no support for
combining the two different ways in the same conference.