Vertically align sponsor invoice address on admin
For addresses spanning multiple rows, it becomes unclear if the
display name spans two rows or if it's the address. To avoid any
confusion around setting the right name for a sponsor, vertical
align the address to the top of the table element.

Setting style on the individual <th> element is less than elegant,
but without a clear id/class hierarchy to target just this element
it's less evil than the option at the moment. Longer term perhaps
all tables should be vertically aligned to top but thats a much
larger patch.