Fix amount sent to Braintree
When the amount already includes a decimal place, we were mistakenly
trying to add another one, causing Braintree to fail.

As Braintree will actually accept either '10' or '10.00', it seems
unnecessary to try and do anything here, so just remove the addition
of the '.00'.

With this, Braintree processing against their sandbox works with the
current version of python-braintree and python-openssl from
jessie-backports.