Introduce send_template_mail
This introduces the function of the same name from the pgweb code base
(though it isn't identical, to be clear).  This function is to be used
when rendering a text file into an email.  There is a helper function
template_to_string() which is also added and can be used if a string
result is needed.

The primary different here is that the variables set in
util/context_processor:settings_context are included in the hash
automatically, so those settings do not need to be explicitly set by
the callers and the templates can expect them to be available.

Using that, this change also includes changes to move away from
hard-coded values like 'PostgreSQL Europe' in favor of using the
variables set in settings.py/local_settings.py in the text templates.
This will allow others to re-use the exisitng templates and not have
to develop and maintain their own going forward.