Correct URL to sponsorship contracts
This URL is only used from /admin/, but should still to to the actual
location of the PDF of course.

This is likely badly broken (still) for things other than sponsorship
contracts per the pre-existing comment in the code, but none of those
use admin views at this point, so that fix will have to be another day.