Add registration specialtype for confirmed or reserve speaker
Previous type only applied to confirmed speakers. In a lot of scenarios,
it makes more sense to allow both confirmed and reserve listed speakers.
In order to keep the flexibility, add this as a separate specialtype to
choose from.

The intention is of course that each conference should only be using one
of these types, but is given the flexibility to choose which one.