Include payment link on PDF invoices
This helps with cases when just the PDF invoice is passed along.
Unfortunately it's still a very long link, so it's not convenient for
typing in if delivered on paper. For this case there is also a QR code
included with it. For online viewing the link directly in the PDF is
clickable and goes to the link that does not require a login.