Add support for EU reverse vat, and change sponsor invoice generation
Per the new instructions from the tax authorities, we should do 100%
opposite of their previous instructions, and instead actually issue
reverse VAT invoices for sponsorship.

To do this, we implement a new flag on each invoice that indicates if it
should have reverse VAT. If this is used, the actual VAT must be zero,
and instead it will trigger text about it on the invoice.

Sponsors inside the EU and with a valid VAT number, but *not* inside
France, will get reverse VAT. Any sponsor without a VAT number *or* a
sponsor in France with a VAT number will get regular VAT. No invoices
are issued without VAT.

Based on this we no longer keep track of VAT number separately from "in
vat area" as our previous instructions required, so drop this field from
the database.

Finally, uses screen scraping to validate VAT numbers on the EU websites
since their SOAP service is basically broken. This can easily be turned
off in the config file (and is turned off by default) in case the
service stops working.