Implement volunteer schedule management
This adds the concepts of VolunteerSlot and VolunteerAssignment.

Each VolunteerSlot has a minimum and a maximum number of volunteers.
Possible volunteers (listed in a new field on the conference itself) can
sign up for specific slots they are interested in, which must then be
confirmed by an admin (controlled by the admin field on the conference).
Admins can also add volunteers to specific slots, which must then be
confirmed by the volunteer.

There is also a per-volunteer ical feed using a secret URL (this adds a
regtoken field to all registrations, which is something we've already
needed for some other things as well, so it'll be good to have
regardless).

No notifications are sent in this system, that's all intended to be
manually handled by the volunteer manager, at least for now.