Make it easier to do partial refunds
In the most common case where the same VAT rate is used everywhere, make
it possible to just type in a percentage and have the system calculate
the actual amounts to be refunded.

Calculation is done client side and the user can/will have to change the
exact amounts for example if it comes out as a non-round number. This is
just a shortcut to avoid having to do manual calculation.