Add functionality to transfer a registration confirmation
This can be done to replace an already confirmed registration with
another one, as a pure substituation. To work, the recipient still has
to register, and just not proceed to generate invoice or sign up on
waitlist.

Will transfer payment information as well as additional options. Will
also send notification both to old and new users. There are most likely
still some corner cases not covered, but even more corner cases are
missed when transferred manually.