Attempt to implement concurrently safe admin
There is an issue with concurrent updates in django admin, in that the
entire entry will be overwritten on save, even if some fields have been
edited concurrently, say by a notification arriving or a cronjob or
something like that.

To avoid this issue, take a hash of all the values when generating a
form and store it in a hidden field. When a form is submitted, compare
this hash to the current values in the database, and if anything has
changed raise a validation error.

Makes no attempt to actually resolve the conflict, it only sets a global
error on the form, and the user has to reload and start over, but it
should prevent concurrent modification.

Has worked in testing, but there are probably some cases where this will
incorrectly reject an appropriate update. Let's hope it's not too many.
There are also likely cases where it misses things, but we should be no
worse off than before this patch.