Render schedule from SQL instead of looping over ORM
This makes it a lot cheaper to render the schedule, and increasingly so
as we add features to it (like the figuring out if there are slides for
a specific talk).

Even more simplification can be done if/when we drop support for the way
we render the schedule in the main website now, where the actual
coordinates are calculated in the python code (and thus very
non-responsive). We should probalby back-port the new responsive
schedule code from a recent conference into the main repo, and can then
get rid of that part as well.