Rewrite cross-conference email system
This mainly adds two features:

* Ability to opt-out of cross conference emails for attendees. Each
  outgoing email will include a link with a token which can be used to
  opt out, at which point only direct conference emails will be
  received.
* A more flexible way to include who to send email to. This includes the
  ability to send to speakers (even those that were not registered, or
  those that were registered with a  different email address). This can
  in the future be extended to also include things like specific
  additional options.

To implement the opt-out feature, a new concept of a "Conference Series"
has been added. This groups the conferences into series like "pgconfeu",
"fosdem pgday" etc, and makes it possible to opt-out on the basis of an
individual series. The series will be used for other things in the
future as well, but for now it's only for opt-out. Of course it is also
possible to opt-out globally.

This adds a "speakertoken" field to all speakers to make it possible to
link to these profiles without a login as well.