Make registration and speaker tokens mandatory fields
Also automatically update all existing records to get a new random
token.

This requires the pgcrypto extension to be installed into the pgcrypto
schema in the database, so that new tokens can be generated. It's only
used in the migration, the existing code is used to generate new tokens
for old entries.