Properly set speaker token on empty speaker creation
When an explicit speaker was created, the token was set properly, but
when it was implicitly created by submitting a new talk without creating
a profile first, the speaker was created with no token.

Unfortunately, django validation only works at the form level, so direct
modifications to the model are untable to validate these things.