The Great URL Rename (TM) project
This commit changes almost every confreg URL in the system, from using
the pattern /events/action/conference to using
/events/conference/action. E.g. /events/register/someconf/ to
/events/someconf/register/.

The reason for this is that all pages that belong to one event now live
under /events/eventname/, making it more predictable and enablinge some
easier changes in the future as well.

Redirects are implemented for all the older URLs (except there is of
course likely something missed somewhere...).