Add ability to copy submissions from a previous conference
For people that submit to many of our events, this will cut down the
copy/paste work.

We make a copy of the submission, so the user can edit eh specifics, and
will make a note in the submission notes field that this was a copy.

For the time being we include all old submissions the speaker has made.
In the future we may want to put some sort of limit on it, and also
bring in some ability for the speaker to see which previous events the
same talk has been submitted to (and for reviewers).