Remove support for non-jinja based conference skinning
The jinja skinning is basically better in every aspect, and by removing
support for the pre-jinja code we no longer have to maintain a whole lot
of duplicated templates.

In exchange for that we now have to maintain a duplicated base.html
template :) Would be good to get that sorted at some point as well, but
that's for later.

This includes dropping the database columns holding the data from
previous conferences, and it means we can get rid of the ugly setup with
symlinks in the directory conference_templates.