Fix githash setting for deploystatic:ed templates
When using deploystatic to deploy directly from a repository branch,
there is no git repository in the resulting template directory. This
means that when the directory is loaded from the main jinja parser, it
does not find a .git, and thus does not set a githash.

To make this work, have deploystatic drop a file called
.deploystatic_githash into the target directory, and have the jinja code
specifically look for this code when trying to find the githash.