Remove support for non-conference speakerprofile viewing
This was the only case where conference was passed as None into the
rendering system, so we can simplify the code a bit and downstream
skinning (used by postgresql.us for example) by removing the ability.

The speaker profile is still shared. It's just mandatory to specify
which conference in the link, to get the proper skinning.