Implement new workflow for registering multiple attendees
In yet another attempt to make it easier for the cases where either
somebody wants to register for multiple colleagues, or just for a single
colleague who cannot register themselves.

This separates out the registrator role from the attendee role in the
database. Regular registrations still work the same way, and sets
registrator and attendee to the same. But it's now also possible to,
using a separate form for each one, make multiple registrations for
other people, by filling out all their details (including email). These
registrations will then automatically generate a bulk invoice.

When the bulk invoice is paid, the registrations attached to it will get
marked as confirmed by the system. They will then also be attached to
the community account of the attendee, if one exists. If one does not
exist, and email is sent to the attendee with a link using the secret
registration token, that makes it possible to attach this registration
to a different account. This means that the user can sign up for an
account (especially easy now that we have oauth logins available) after
the registration has been created, or they can attach it to a different
account if they already have one but it has a different email address
than used on the registration.