Change title of bulk invoices
Instead of always saying "bulk invoice", say either "<conf>
registration" or "<conf> multiple registrations" depending on how many
registrations are on it. With the new register for somebody else
workflow it's possible to have a a bulk payment with just one
registration on it, so try to be more clear.

Also change the pages viewing a bulk payment when coming in through the
multireg workflow, to not use that word. Since the old way to do bulk
payments is still around, keep the page titles there.