Don't attempt to add errors to a form that's empty
The add_error method of a form will fail if is_valid() has not been
called. So only add the errors in this case -- which means that finalize
was picked. Since this is just the invoice form it's not an issue (it's
a catch-all error path which will only be called if something passed
verification earlier but is now no longer valid, such as an expired
discount code)