Properly flag vouchercodes as unused on canceled bulk invoices
This was very uncommon before, but now that we use bulk invoices for
multi-reg processing it becomes more common, and this case shows up more
often. We need to mark the vouchercodes as unused in the system,
otherwise we will fail to generate a new invoice for the same
registration once one has been canceled.