Fix detection of pending multiregs
The correct detection of pending multiregs is required to show the
"Create invoice" button to users. However, the existing code would look
at the first registration in the list (with no order by) and check if it
was pending, never bothering to look at the rest. Instead, it's supposed
to look at all of them, and indicate true if *any* pending registrations
exist.