Add functionality to send email to people based on signup
This makes it possible both to send to people who have been added to a
signup (for exampe, to announce the signup), but also to people based on
which selections they made in the signup (e.g. yes or no).

It's only possible to pick one specific answer, so if one wants to cover
multiple answers (such as "yes", and "yes with a +1") emails simply have
to be sent more than once.

Emails are delivered as-is, and a note sent to conference organizers. In
particular, they do *not* show up on the dashboard under sent emails.
For things to show up there, the regular email sending must be done.