Don't flag vouchers/discounts as used until multireg completed
If the invoice details are not filled in, we generate an error in the
form. But at the time we had already flagged the voucher/code as used,
and thus the next create would now fail saying the code has already been
used.

Fix this by setting a savepoint at the beginning and roll back to it in
the failure case.