Support using google charts trendlines in time reports
Instead of just assuming that nothing at all will happen from the
current day, use the trendline functionality in google charts to map it.

Trendlines will be applied to only the latest of the conferences if more
tha none is picked -- this seems to be what google supports.

Trendlines are only supported for multiconference reports, mostly for
this reason.