Move prepaid batch handling into the admin framework
This includes both the HTML framework like most previous commits, but
also management.

A conference admin (non-superuser) can now view prepaid batches, since
it no longer uses /admin/. They can also create new prepaid batches with
vouchers as necessary (though of course only for their own conferences).

This also removes the last legacy connection to the previous invoice
management system that was in confreg, namely relying on the invoice
management permission from that system.

And finally, it changes the creation form to use selectize and the
django plugin to do the search-for-users-when-the-list-is-too-long
functionality, so it uses the same as everything else.