Lay the groundwork for moving metadata into the backend system
Previously all metadata was edited throuhg /admin/, which wan't very
user friendly, gave no way to separate out "normal things" and "internal
things" when editing, and made it hard to do permissions checking.

This commit creates the new concept of a "backend list" and "backend
form", which can be used as shortcuts to implement editors in the normal
backend framework instead of the admin framework. These editors are
based off a regular Form (inherited from BackendForm), and can create
all things automatically for all simple cases. All access is
automatically authenticated to conference administrators for the
specific conference.

The term backend was picked specifically to avoid name-confusion with
the builtin django admin module.

Finally, to prove the functionality, a few forms and lists were
implemented as part of the original step:

* Editor for the conference metadata itself (limited fields)
* Editor for a registration (limited fields)
* List and editor for registration classes
* List and editor for registration types

There are many more to come there before things are finished, but it's a
starting point.