Fix JSON/badge report when there are regs with no regtype
This could only occur for attendees that aren't actually confirmed, so
normally they would not be included in such a report because of a
filter. However, just "dump everything" is a pretty common use-case for
testing, and obviously we shouldn't *crash* in this scenario.

Spotted by ads, when doing exactly that.