Attempt to better handle payment exceptions in adyen processing
The invoice API actually allows us to get the reason for why the
processing failed, so use that instead of just throwing an exception.

In particular this lets us track the case of invalid size payment in
relation to invoice amount (never happened so far), and also payments
that happen after the invoice has been deleted (theoretically possible
with creditcard payments, but very much realistic with iban payments --
it's more surprising it hasn't happened before).

In these cases create an empty accounting record that can be processed,
instead of leaving things in incomplete state.