Add support for copying data from previous conference
This will make it easier to copy things like feedback, schdule slots
etc. Uses the backend list functionality already inplace, and supports
basic transformation of the data (such as adding days to a timeslot so
that it falls within the new conference days). Each different type of
object needs to implement the actual copying functionality, to make sure
it can deal with mappings and such.