Add tool to flag completed Trustly refunds
For Trustly we don't get separate notificatoins when refunds complete.
So instead, add this tool that polls the status of any pending refunds,
and completes the process when they are listed as refunded.

This has been missing functionality since we added Trustly payments, but
we haven't had any refunds to process yet..