Don't ask all registrants for phone numbers
We don't really need it, so we shouldn't ask for it.

Instead, make it possible to ask for it on a per-registration-type
basis, and make it mandatory on that. This means we can now say that for
example Staff members and Speakers need to provide a phone number, but
not attendees.