Add extra fields to all VAT queries
It seems some countries, such as Spain, will fail in VAT validation if a
set of extra fields are not included. These fields can be set to blank
and it works -- and it seems to work fine to include them for other
countries as well, so let's include them in all requests and set them to empty.