Use actual URL field for the sponsor signup URLs
This enables both client side validation that what's put in there is an
URL, and server side verification of that.

We don't ping back and check the URL, because that could be used to do
evil things to say APIs that check source IP. We may want to look into
doing that securely in the future, but this is a good starting point.