Implement proxy voting in meeting-votes
A proxy voter is given a name and a unique proxy voting key. This voting
key (generated the same way as our other "secure tokens") can be used to
access the voting for this member, and then return the same access to
the meeting.

A proxy voter can be assigned up to 4 hours before the meeting, this
being the same time as when a regular member can log in to access their
meetings.

Including review/fixes from Dave