Add support for token based access to admin information
This adds a new object for each conference that is an AccessToken. This
token can be given permissions to access specific types of data in a
structured format, which can then be loaded using an URL with that token
in it.

Initially this exports a number of datasets that are useful to feed into
a budget spreadsheet or similar, but the model and code is structured to
make it easy to add completely different exports as well in the future.