Allow editing certain fields on a registration
We don't allow editing the "main" fields on a registration, but allow
things like t-shirt size, dietary needs etc. In particular useful if the
registration is made by somebody else, this allows the actual attendee
to update such fields.

Add a checkbox at the conference level to stop this -- intended to turn
it off once badges are printed for example.