Replace empty paragraph with linebreak
While not a violation of the specification, an empty paragraph is
not recommended in HTML5 and browsers are to discard it. Rendering
as whitespace is only supported in IE7-9, with Chrome and Firefox
dropping it. The intention was most likely to induce some vertical
whitespace so replace with a linebreak tag instead (which indeed
makes the page nicer to look at).