Add support for linked objects in backend forms
This is things that are "sub-objects" of another object. In the admin
interface we typically used inline forms for it, but it will be easier
to use as simple linked objects.

Recursively call into the existing backend form handler to manage them,
as all the code is already in place.