Make selectize fields in backend load entries dynamically
Poplating three dropdowns with every single user on the system becomes a
bit inefficient (to the point that the form in production started
approaching 1Mb in size and took a lot of time to load).

So instead of doing that, make the selectize fields load their contents
dynamically through an Ajax call. To do this, implement three lookups
for accounts, registrations and speakers.