Basic documentation for confreg system
This adds basic contextual aware documentation to the confreg system
(sponsorship system not yet included, but will be at some point in the
future). A help link is added to the top right part of the navigation
bar, which links to (potentially context aware) help pages.

Docs live in markdown format in the repo and are rendered as django
views, inside the framework. Graphs that are used are generated using
graphviz, but checked into the repo both as .dot and as .svg to not make
graphviz a dependency for anything other than actually changing the
graphs. The docs view in django basically does the markdown rendering
straight up but also knows how to inline the svg graphs specifically.