Make it possible to cancel a waitlist offer
Previously it was only possible to remove a user from the waitlist when
they did not have an active offer, and one had to wait until the offer
expired to to that. Now we can cancel the offer but leave them on the
waitlist, and then decide if we want to cancel the waitlist completely
as a second step.