Make it possible to make waitlist offers until a certain time
Previously it was only possible to say "<n>" hours, which made the
administrator do the calculation to make sure the offer didn't expire in
the middle of the night etc.

With this it's now possible to either extend an offer for a certain
number of hours, or until a certain point in time.

In both cases, the fields default to 48 hours from the current time.