Wrap sessionlink in div element to allow styling
The "Back" link on session and speaker display currently takes the
user to the sessionlist, which may not be where the user actually
came from. This wraps the link in a div element such that an event
can hide it with visibility settings in the override CSS in case it
doesn't make sense.