Widen buttons on registration dashboard
The buttons dooesn't fit the text on them, and since their styling
was inline with !important it couldn't be fixed by the conference
overrides. Fix by widening the button and removing !important. This
lipsticking the pig to handle the current event taking place, all
the inline CSS needs to removed at some point but thats another
patch.