Clarify wording on notification queue draining
While it's true that the speakers of the sessions will be attendees
of the conference, they probably won't be at this point. I had to
read code to ensure myself this wasn't sending a notification to
all registered attendees of the conference (which it doesn't), so
to spare myself the same excercise the next time reword the text
to make it clear.