New base design for www.postgresql.eu website
This replaces the base design with something that's a lot more modern,
based on bootstrap etc. Lots of inspiration from the updated main
postgresql.org website. Also is at least attempting to be a little bit
responsive, unlike previous website.

Also restructures the event section completely, to be able to more focus
on the events that we run.

Does not change anything fundamental in the backend systems, and
*should* have no effect on downstream conferences as those templates are
just wrapped in a different otuside now.