Fix cross conference emailing for registered-by-others users
When a user is registrered by somebody else, they get an attendee_id of
NULL. These registrations got filtered out against the global opt-out
list, due to the NULL check.

Make it so that any registration with NULL attendee is always included.
If this user decides to opt out globally, the attendee_id field will get
populated when they do that, and the regular opt-out rules will apply.

Per-conference-series optout not affected as it was already doing the
correct thing.