Don't limit news JSON feed to RSS feed items only
If we limit it like that then there is no way to actually get to those
news items. They should be included in the JSON feed, just not RSS,
because the JSON feed is intended for internal consumption.

Instead, add a boolean field to the actual *contents* representing the
flag for inrss.