Implement a tweet queue for conferences along with time windows
The tweet queue allows different portions of the system to queue up
things that should be posted by the conference twitter.

Also add fields for each conference that define the earliest and latest
in a day that automated tweets will be posted. If a tweet is queued
outside those times, it will stay in the queue and get posted once the
interval is entered.

For each invokation of the script, one tweet per conference is posted.
If more are queued up (or if they are in the news), they remain in the
queue and will be delivered on the next invokation of the script. With
the script executed every 5 or 10 minutes, that ensures a spread of the
tweets.