Make it possible to pass instancemaker through backend lists
This makes it possible to create objects that do not have a field for
conference. It was pure luck that this worked for ConferenceSeries(),
since it actually accepts a conference value (because of the link), but
it shouldn't have worked. So convert ConferenceSeries to use a dedicated
instancemaker as well.