Make it easier to cancel off waitlist once an offer is received
Make a link directly from the registration page to handle this instead
of having to first try to finish and then cancel upon failure.

In passing also update the waitlist offer email to explicitly ask the
recipient to go cancel their entry on the waitlist if they don't
want/need it anymore.