Add a step for previewing sponsor invoice addresses
Due to the number of sponsors that got it wrong, make a separate step
where they get to proview the sponsor address exactly as it's going to
look on the invoice before proceeding. It won't help with the ones who
actually put the wrong company name in, but at least for things like
VAT numbers and duplicated (almost the same but not quite) company names
etc.