Require administrator approval for isntantbuy sponsorships
As these would immediately proceed to invoice stage before, they could
lead to a whole lot of extra work if incorrect details were submitted.
To avoid this, require an administrator click-to-approve step before the
invoice is generated, which should make it easier to abort-and-retry in
case there is an error in the signup.