Make it possible to reject a sponsorship
Before a sponsorship is confirmed, it can now be rejected instead of
being confirmed. When rejecting a sponsorship, a reason must be given
and this one is sent to the person who signed up.

In passing, change the "generate invoice" button to be on a form and do
a POSTback instead of a GET.