Handle Adyen settlement batches with just balancetransfers
If we have a batch with just balance transfers going in and out of
exactly the same amount, an exception would be thrown because they'd be
summarized into 0.

Fix that by splitting apart incoming and outgoing balance transfers,
making each one of them non-zero.

For most normal settlement batches this was not a problem since
they will only have one Balancetranfer row, and it will be balanced
against the other types of entries.