Reimplement multiple speaker handling in call for papers
Instead of the awkward way with a subform, we now use selectize to
populate a "dynamic" list of speakers. This is the same technology we've
arleady used in the backend for some time with good results.

This adds a dependency on jquery to the conference templates. If a
conference template does not include jquery, multiple speakers will not
work (but the rest of the form will keep working fine). This should be
included before the pagescript block in the template to make sure it's
available when the page runs.